requirements.txt
# Typing

# Django
django==5.2
django-cors-headers==4.7.0

# Two-factor authentication

# Django REST
djangorestframework==3.16.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.94.1
djangorestframework-camel-case==1.4.2
django-filter==23.5

# Celery

# OpenAI

# Miscellaneous
python-dotenv==1.0.1
setuptools==78.1.0

# DB
psycopg2-binary